requirements.txt
beautifulsoup4==4.12.2
blurhash==1.1.4
certifi==2023.5.7
charset-normalizer==3.1.0
datefinder==0.7.3
decorator==5.1.1
idna==3.4
Mastodon.py==1.8.1
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2023.3
regex==2023.5.5
requests==2.30.0
six==1.16.0
soupsieve==2.4.1
urllib3==2.0.2
cloudscraper==1.2.71
